Add DCParametersSPA validation model
